requirements.txt
Django==1.3.1
-e hg+http://bitbucket.org/jesselegg/aashe-python#egg=aashe-python
markdown
django-treemenus
BeautifulSoup
Sphinx
Fabric
pyPdf
requests
django-csv-importer